chat, chatRoom controller, service class 분리
