+++ src/main/java/com/example/weup/controller/ChatController.java
@@ -24,59 +24,8 @@ import java.util.List;
 public class ChatController {
 
     private final ChatService chatService;
+
     private final SimpMessagingTemplate messagingTemplate;
-
-    @ResponseBody
-    @PostMapping("/chat/create")
-    public ResponseEntity<DataResponseDTO<String>> createChatRoom(@LoginUser User user,
-                                                                  @RequestBody CreateChatRoomDTO createChatRoomDto) {
-
-        chatService.createChatRoom(user, createChatRoomDto);
-
-        return ResponseEntity.ok(DataResponseDTO.of("채팅방 생성이 완료되었습니다."));
-    }
-
-    @ResponseBody
-    @PostMapping("/chat/list/{chatRoomId}")
-    public ResponseEntity<DataResponseDTO<List<GetInvitableListDTO>>> getMemberNotInChatRoom (@LoginUser User user,
-                                                                                              @PathVariable Long chatRoomId) {
-
-        List<GetInvitableListDTO> getInvitableList = chatService.getMemberNotInChatRoom(chatRoomId);
-
-        return ResponseEntity.ok(DataResponseDTO.of(getInvitableList, "채팅방 초대 가능한 멤버를 불러왔습니다."));
-    }
-
-    @ResponseBody
-    @PostMapping("/chat/invite/{chatRoomId}")
-    public ResponseEntity<DataResponseDTO<String>> inviteChatRoom(@LoginUser User user,
-                                                                  @PathVariable Long chatRoomId,
-                                                                  @RequestBody InviteChatRoomDTO inviteChatRoomDTO) throws JsonProcessingException {
-
-        chatService.inviteChatMember(chatRoomId, inviteChatRoomDTO);
-
-        return ResponseEntity.ok(DataResponseDTO.of("채팅방에 초대되었습니다."));
-    }
-
-    @ResponseBody
-    @PostMapping("/chat/edit/{chatRoomId}")
-    public ResponseEntity<DataResponseDTO<String>> editChatRoom(@LoginUser User user,
-                                                                @PathVariable Long chatRoomId,
-                                                                @RequestBody String chatRoomName) {
-
-        chatService.editChatRoomName(chatRoomId, chatRoomName);
-
-        return ResponseEntity.ok(DataResponseDTO.of("채팅방 이름이 수정되었습니다."));
-    }
-
-    @ResponseBody
-    @PostMapping("/chat/list/{projectId}")
-    public ResponseEntity<DataResponseDTO<List<GetChatRoomListDTO>>> listChatRoom(@LoginUser User user,
-                                                                             @PathVariable Long projectId) {
-
-        List<GetChatRoomListDTO> data = chatService.getChatRoomList(user, projectId);
-
-        return ResponseEntity.ok(DataResponseDTO.of(data, "채팅방 리스트를 불러왔습니다."));
-    }
 
     @MessageMapping("/send/{chatRoomId}")
     public void sendMessage(@DestinationVariable Long chatRoomId, SendMessageRequestDTO messageDto) throws JsonProcessingException {
@@ -113,13 +62,4 @@ public class ChatController {
         return ResponseEntity.ok(DataResponseDTO.of(data, "채팅 내역 조회가 완료되었습니다."));
     }
 
-    @ResponseBody
-    @DeleteMapping("/chat/leave/{chatRoomId}")
-    public ResponseEntity<DataResponseDTO<String>> exitChatRoom(@LoginUser User user,
-                                                                @PathVariable Long chatRoomId) throws JsonProcessingException {
-
-        chatService.leaveChatRoom(user, chatRoomId);
-
-        return ResponseEntity.ok(DataResponseDTO.of("채팅방에서 퇴장하였습니다."));
-    }
 }
