+++ src/main/java/com/example/weup/controller/ProjectController.java
@@ -8,10 +8,9 @@ import com.example.weup.dto.response.DetailProjectResponseDTO;
 import com.example.weup.dto.response.ListUpProjectResponseDTO;
 import com.example.weup.dto.response.ResponseDTO;
 import com.example.weup.entity.ChatRoom;
-import com.example.weup.entity.ChatRoomMember;
 import com.example.weup.entity.Member;
 import com.example.weup.entity.Project;
-import com.example.weup.service.ChatService;
+import com.example.weup.service.ChatRoomService;
 import com.example.weup.service.MemberService;
 import com.example.weup.service.ProjectService;
 import jakarta.validation.Valid;
@@ -32,7 +31,8 @@ public class ProjectController {
     private final ProjectService projectService;
 
     private final MemberService memberService;
-    private final ChatService chatService;
+
+    private final ChatRoomService chatRoomService;
 
     @PostMapping("/create")
     public ResponseEntity<ResponseDTO> createProject(@LoginUser Long userId,
@@ -41,8 +41,8 @@ public class ProjectController {
         log.info("요청자 : {}, create project -> start", userId);
         Project newProject = projectService.createProject(projectCreateRequestDTO);
         Member newMember = memberService.addProjectCreater(userId, newProject);
-        ChatRoom newChatRoom = chatService.createBasicChatRoom(newProject, projectCreateRequestDTO.getProjectName());
-        chatService.addChatRoomMember(newProject, newChatRoom, newMember.getMemberId());
+        ChatRoom newChatRoom = chatRoomService.createBasicChatRoom(newProject, projectCreateRequestDTO.getProjectName());
+        chatRoomService.addChatRoomMember(newProject, newChatRoom, newMember.getMemberId());
 
         log.info("요청자 : {}, create project -> success", userId);
         return ResponseEntity.ok(DataResponseDTO.of("프로젝트가 성공적으로 생성되었습니다."));
