+++ src/main/java/com/example/weup/repository/ChatMessageRepository.java
@@ -3,7 +3,6 @@ package com.example.weup.repository;
 import com.example.weup.entity.ChatMessage;
 import com.example.weup.entity.ChatRoom;
 import com.example.weup.entity.Member;
-import com.example.weup.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -16,9 +15,9 @@ public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long> 
 
     void deleteByChatRoom(ChatRoom chatRoom);
 
-    List<ChatMessage> findByUser(User user);
+    //List<ChatMessage> findByUser(User user);
 
     ChatMessage findTopByChatRoom_ChatRoomIdOrderBySentAtDesc(Long roomId);
 
-    List<ChatMessage> findBySenderId(Member member);
+    List<ChatMessage> findByMember(Member member);
 }
